express runs as ints

+++ metrics.py
@@ -27,12 +27,12 @@ class Metrics:
 		away_max, home_max = max(away_inning_runs), max(home_inning_runs)
 
 		result = {
-			'away_max' : away_max,
-			'away_inning' : away_inning_runs.index(away_max) + 1,
-			'away_value' : away_max * away_runs,
-			'home_max' : home_max,
-			'home_inning' : home_inning_runs.index(home_max) + 1,
-			'home_value' : home_max * home_runs
+			'away_max' : int(away_max),
+			'away_inning' : int(away_inning_runs.index(away_max)) + 1,
+			'away_value' : int(away_max * away_runs),
+			'home_max' : int(home_max),
+			'home_inning' : int(home_inning_runs.index(home_max)) + 1,
+			'home_value' : int(home_max * home_runs)
 		}
 		return result
